backend/requirements.txt
fastapi
uvicorn
pydantic[email]
python-dotenv
PyJWT
sqlmodel
asyncpg
psycopg2-binary
bcrypt
python-multipart
httpx
pika
